test: organize extension entry tests
