+++ common/extension/config_loader_test.go
@@ -25,7 +25,9 @@ import (
 import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
+)
 
+import (
 	"dubbo.apache.org/dubbo-go/v3/filter"
 )
 
@@ -146,5 +148,5 @@ func TestInitializeIgnoresOtherRoleYAML(t *testing.T) {
 
 func TestMergeFilterNamesHonorsExplicitSuppression(t *testing.T) {
 	assert.Equal(t, "-extension,a,b", MergeFilterNames("-extension,a,a", []string{"extension", "b", "b"}))
-	assert.Equal(t, "", MergeFilterNames("", nil))
+	assert.Empty(t, MergeFilterNames("", nil))
 }
